+++ metro-property-management/src/Components/Home.jsx
@@ -1,15 +1,15 @@
 import React from "react";
 import { Link } from "react-router-dom";
 
-import Navbar from "../Header/Navbar";
-import HeroSection from "../HeroSection/HeroSection";
-import TeamSection from "../TeamSection/TeamSection";
-import Footer from "../Footer/Footer";
-import ContactSection from "../ContactSection/ContactSection";
-import ChooseUsSection from "../ChooseUsSection/ChooseUsSection";
-import TenantAppSection from "../TenantAppSection/TenantAppSection";
-import SucessSection from "../SucessSection/SucessSection";
-import styles from "../Home.module.css";
+import Navbar from "./Header/Navbar";
+import HeroSection from "./HeroSection/HeroSection";
+import TeamSection from "./TeamSection/TeamSection";
+import Footer from "./Footer/Footer";
+import ContactSection from "./ContactSection/ContactSection";
+import ChooseUsSection from "./ChooseUsSection/ChooseUsSection";
+import TenantAppSection from "./TenantAppSection/TenantAppSection";
+import SucessSection from "./SucessSection/SucessSection";
+import styles from "./Home.module.css";
 
 function Home() {
   return (
